+++ src/counterfactuals/Counterfactuals.jl
@@ -4,7 +4,7 @@ using .GenerativeModels
 using .Generators
 using .Models
 using ChainRulesCore: ChainRulesCore
-using Flux
+using Flux: Flux
 using MLUtils: MLUtils
 using MultivariateStats
 using Statistics: Statistics
@@ -14,7 +14,6 @@ using StatsBase
 include("core_struct.jl")
 include("encodings.jl")
 include("generate_counterfactual.jl")
-include("growing_spheres.jl")
 include("info_extraction.jl")
 include("initialisation.jl")
 include("path_tracking.jl")
@@ -23,6 +22,7 @@ include("search.jl")
 include("termination.jl")
 include("utils.jl")
 include("vectorised.jl")
+include("flatten.jl")
 
 # Counterfactual Rule Explanations:
 include("CRE.jl")