+++ src/evaluation/evaluate.jl
@@ -1,6 +1,43 @@
 using DataFrames: nrow
 using UUIDs: uuid1
 
+"An abstract type for CE transformers."
+abstract type AbstractCETransformer end
+
+"Default CE transformer that returns the input as is."
+struct IdentityTransformer <: AbstractCETransformer end
+
+"Transformation function for default transformer."
+TransformationFunction(transformer::IdentityTransformer) = (x -> x)
+
+"Global CE transformer."
+global _ce_transform = TransformationFunction(IdentityTransformer())
+
+"Get the global CE transformer."
+get_global_ce_transform() = _ce_transform
+
+"The `ExplicitCETransformer` can be used to specify any arbitrary CE transformation."
+struct ExplicitCETransformer <: AbstractCETransformer
+    fun::Function
+    function ExplicitCETransformer(fun::Function)
+        @assert hasmethod(fun, Tuple{CounterfactualExplanation}) "Measure function must have a method for `CounterfactualExplanation`"
+        return new(fun)
+    end
+end
+
+"Transformation function for explicit transformer."
+TransformationFunction(transformer::ExplicitCETransformer) = transformer.fun
+
+"""
+    global_ce_transform(transformer::AbstractCETransformer)
+
+Sets the global CE transformer to `transformer`.
+"""
+function global_ce_transform(transformer::AbstractCETransformer)
+    global _ce_transform = TransformationFunction(transformer)
+    return _ce_transform
+end
+
 """
     compute_measure(ce::CounterfactualExplanation, measure::Function, agg::Function)
 
@@ -24,14 +61,15 @@ function to_dict(computed_measures::Vector, measure)
 end
 
 """
-    evaluate_dataframe(
-        ce::CounterfactualExplanation,
-        measure::Vector{Function},
-        agg::Function,
+    to_dataframe(
+        computed_measures::Vector,
+        measure,
         report_each::Bool,
         pivot_longer::Bool,
         store_ce::Bool,
+        ce::CounterfactualExplanation,
     )
+
 Evaluates a counterfactual explanation and returns a dataframe of evaluation measures.
 """
 function to_dataframe(
@@ -53,14 +91,20 @@ function to_dataframe(
         evaluation = DataFrames.stack(evaluation, DataFrames.Not(:num_counterfactual))
     end
     if store_ce
-        evaluation.ce = repeat([ce], nrow(evaluation))
+        transform_fun = get_global_ce_transform()
+        evaluation.ce = repeat([transform_fun(ce)], nrow(evaluation))
     end
     DataFrames.select!(evaluation, :num_counterfactual, :)
     return evaluation
 end
 
 """
-    generate_meta_data(i::Int, ce::CounterfactualExplanation, evaluation::DataFrame, report_meta::Bool, meta_data::Union{Nothing,Vector{Dict}})
+    generate_meta_data(
+        ce::CounterfactualExplanation,
+        evaluation::DataFrames.DataFrame,
+        meta_data::Union{Nothing,Dict},
+    )
+
 Generates meta data for a counterfactual explanation. If `report_meta=true`, the meta data is extracted from the counterfactual explanation. If `meta_data` is supplied, it is used instead.
 """
 
@@ -82,15 +126,31 @@ end
 
 """
     evaluate(
-        ce::CounterfactualExplanation;
+        ce::CounterfactualExplanation,
+        meta_data::Union{Nothing,Dict}=nothing;
         measure::Union{Function,Vector{Function}}=default_measures,
         agg::Function=mean,
         report_each::Bool=false,
         output_format::Symbol=:Vector,
-        pivot_longer::Bool=true
+        pivot_longer::Bool=true,
+        store_ce::Bool=false,
+        report_meta::Bool=false,
     )
 
 Just computes evaluation `measures` for the counterfactual explanation. By default, no meta data is reported. For `report_meta=true`, meta data is automatically inferred, unless this overwritten by `meta_data`. The optional `meta_data` argument should be a vector of dictionaries of the same length as the vector of counterfactual explanations. 
+
+## Arguments:
+
+- `ce`: The counterfactual explanation to evaluate.
+- `meta_data`: A vector of dictionaries containing meta data for each counterfactual explanation. If not provided, the default meta data is inferred from the counterfactual explanations.
+- `measure`: The evaluation measures to compute. By default, all available measures are computed.
+- `agg`: The aggregation function to use for the evaluation measures. By default, the mean is used.
+- `report_each`: If true, each evaluation measure is reported separately. Otherwise, the mean of all measures is reported.
+- `output_format`: The format of the output. By default, a vector is returned.
+- `pivot_longer`: If true, the evaluation measures are pivoted longer. Otherwise, they are stacked.
+- `store_ce`: If true, the counterfactual explanation is stored in the evaluation DataFrame. **Note**: These objects are potentially large and can consume a lot of memory.
+- `report_meta`: If true, meta data is reported. Otherwise, it is not.
+
 """
 function evaluate(
     ce::CounterfactualExplanation,