+++ src/generators/Generators.jl
@@ -5,7 +5,7 @@ module Generators
 
 using ..CounterfactualExplanations
 using ..GenerativeModels
-using Flux
+using Flux: Flux
 using LinearAlgebra
 using ..Models
 using ..Convergence
@@ -26,7 +26,6 @@ export FeatureTweakGenerator
 export GenericGenerator
 export GravitationalGenerator
 export GreedyGenerator
-export GrowingSpheresGenerator
 export REVISEGenerator
 export WachterGenerator
 export FeatureTweakGenerator
@@ -63,7 +62,6 @@ generator_catalogue = Dict(
     :generic => Generators.GenericGenerator,
     :gravitational => Generators.GravitationalGenerator,
     :greedy => Generators.GreedyGenerator,
-    :growing_spheres => Generators.GrowingSpheresGenerator,
     :revise => Generators.REVISEGenerator,
     :dice => Generators.DiCEGenerator,
     :wachter => Generators.WachterGenerator,