+++ src/objectives/Objectives.jl
@@ -1,7 +1,7 @@
 module Objectives
 
 using ..CounterfactualExplanations
-using Flux
+using Flux: Flux
 using Flux.Losses: Losses, logitbinarycrossentropy, logitcrossentropy, mse
 using ChainRulesCore: ChainRulesCore
 using LinearAlgebra
@@ -16,6 +16,7 @@ include("traits.jl")
 export logitbinarycrossentropy, logitcrossentropy, mse, predictive_entropy
 export losses_catalogue
 export distance, distance_mad, distance_l0, distance_l1, distance_l2, distance_linf
+export distance_cosine, distance_from_target, distance_from_target_cosine
 export ddp_diversity
 export EnergyDifferential
 export hinge_loss
@@ -33,7 +34,10 @@ const penalties_catalogue = Dict(
     :distance_l1 => distance_l1,
     :distance_l2 => distance_l2,
     :distance_linf => distance_linf,
+    :distance_cosine => distance_cosine,
     :ddp_diversity => ddp_diversity,
+    :distance_from_target => distance_from_target,
+    :distance_from_target_cosine => distance_from_target_cosine,
     :energy_constraint => energy_constraint,
     :energy_differential => EnergyDifferential(),
     :hinge_loss => hinge_loss,